fixing errors

+++ app_modules/main/OfferDetails.js
@@ -2,6 +2,7 @@ import {obj as ajax } from 'utilities/xhr/ajax';
 import { EditOffer }  from "main/EditOffer";
 import { Dialog } from "utilities/Dialog/dialog";
 import { commonService } from "utilities/common/commonService";
+
 class OfferDetails{
 	constructor(offer){
 		this.dom=null;
@@ -13,6 +14,11 @@ class OfferDetails{
 				request.onload = () => {
 					this.dom = document.createElement("div");
 					this.dom.innerHTML = request.responseText;
+					let userId=localStorage.getItem("userName");
+					if(userId!=this.offer.createdBy){
+						var child=this.dom.querySelector(".edit-delete");
+						this.dom.removeChild(child);
+					}
 					  resolve();
 				};
 		        request.onerror = () => {
@@ -24,43 +30,46 @@ class OfferDetails{
 	}
 	
 	initializeEventHanders(){
-		var deleteNode=this.dom.querySelector("#offer_delete");
-		deleteNode.addEventListener("click",()=>{
-			var canDelete=confirm("do you want to delete the offer?");
-			if(canDelete){
-				ajax.deleteOffer(this.offer).then((res)=>{
-								if(res=="success"){
-									let event = new Event("close-dialog");
-									document.dispatchEvent(event);
-									commonService.showOffers();
-								}else{
-								alert("error while deleting offer");
-								}  
-					},(error)=>{
-						var err=JSON.parse(error);
-						if(err.status && err.status==401){
-							alert("session expired");
-							commonService.clearScreen();
+		let userId=localStorage.getItem("userName");
+		if(userId==this.offer.createdBy){
+					var deleteNode=this.dom.querySelector("#offer_delete");
+					deleteNode.addEventListener("click",()=>{
+						var canDelete=confirm("do you want to delete the offer?");
+						if(canDelete){
+							ajax.deleteOffer(this.offer).then((res)=>{
+											if(res=="success"){
+												let event = new Event("close-dialog");
+												document.dispatchEvent(event);
+												commonService.showOffers();
+											}else{
+											alert("error while deleting offer");
+											}  
+								},(error)=>{
+									var err=JSON.parse(error);
+									if(err.status && err.status==401){
+										alert("session expired");
+										commonService.clearScreen();
+									}
+								});
 						}
 					});
-			}
-		});
-		var editNode=this.dom.querySelector("#offer_edit");
-		editNode.addEventListener("click",()=>{
-			let event = new Event("close-dialog");
-			document.dispatchEvent(event);
-			let dialog = new Dialog({
-				title: "Edit Offer"
-			  });
-			let editOfferWidget= new EditOffer(this.offer);
-			  editOfferWidget.renderPage().then(()=>{
-				dialog.init(editOfferWidget.dom);
-				editOfferWidget.initializeHandlers();
-				editOfferWidget.popUpDetails();
-			  },()=>{
-				console.log("error while creating login widget")
-			  });    
-		});
+					var editNode=this.dom.querySelector("#offer_edit");
+					editNode.addEventListener("click",()=>{
+						let event = new Event("close-dialog");
+						document.dispatchEvent(event);
+						let dialog = new Dialog({
+							title: "Edit Offer"
+							});
+						let editOfferWidget= new EditOffer(this.offer);
+							editOfferWidget.renderPage().then(()=>{
+							dialog.init(editOfferWidget.dom);
+							editOfferWidget.initializeHandlers();
+							editOfferWidget.popUpDetails();
+							},()=>{
+							console.log("error while creating login widget")
+							});    
+					});
+		}
 	}
 	
 	populateOfferDetails(){
